Get rid of `temp_note_value`
After examination, it seems that there is no need for separate buffering,
and directly writing to the period member works the same, but saves bytes
and cycles by cutting the middleman.

Toneporta throws a wrench into this, though, since it requires the
requested note not to actually be played back. This is taken care of by
what's arguably a bodge, but I believe this change is still overall a win.